Deleting an element of an array
program to delete an element of an array of the position of the user's choice.
In this program, the user is asked to enter the position of the element he/she wants to delete but we can also make a program in which the user will only tell the particular element for his/her ease. That would require a search feature in the above code.